requirements.txt
opencv-python
pyzbar
validators
ndeflib